+++ Dojo/Statistics.cs
@@ -1,0 +1,24 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Security.Principal;
+using System.Threading.Tasks;
+using Xunit;
+
+namespace StatisticsDojo
+{
+    public class Statiscis
+    {
+        
+    }
+
+    class Evaluator
+    {
+        public Evaluator()
+        {
+
+        }
+        public int MinValue { get; set; }
+        public int MaxValue { get; set; }
+    }
+}
